+++ NEWS.md
@@ -1,6 +1,7 @@
 # mire (development version)
 
 * Changed `unfill()` to accept a vector of `values` instead of a scalar `value`.
+* Added new function `matpow()` and infix operator `%^%` for powers of a matrix.
 
 # mire 0.1.0 (2023-10-29)